requirements.txt
googletrans==3.0.0
gTTS==2.3.2
matplotlib==3.5.2
pandas==1.4.4
Pillow==9.2.0
Pillow==10.0.0
pytube==12.1.3
Requests==2.31.0
streamlit==1.22.0
streamlit_option_menu==0.3.2
